Add fallback support for Node 17 and make it easier to test the asset
We have updated the readme and added try/catch logic in our node sample so that it works with Node 17, 18, 19.

We have also updated the first manifest sample so that the asset will always be fetched regardless of version (you shouldn't use that setting in production but it is handy for development). We added a way of launching the second manifest and updated it's app asset settings. Launching the second manifest means you have to change the asset version number in order to fetch the latest version of the asset.